Major refactoring to finally implement the CKMS datastream algorithm. Added some flexibility to allow users of the library to inject their own implementation of Stream. Other things...
