+++ simpleclient/src/main/java/io/prometheus/client/Invariant.java
@@ -1,0 +1,8 @@
+package io.prometheus.client;
+
+/**
+ * Created by warebot on 1/4/16.
+ */
+public abstract class Invariant {
+    public abstract double f(double r, int n);
+}
