+++ simpleclient/src/main/java/io/prometheus/client/Quantile.java
@@ -3,16 +3,18 @@ package io.prometheus.client;
 /**
  * Created by warebot on 12/27/15.
  */
+
 /**
  * <p>
- * Quantile is an invariant for estimation with {@link StreamCopy}.
+ * Quantile is an invariant for estimation with {@link Stream}.
  * </p>
  */
+
+
 public class Quantile {
     final double quantile;
     final double error;
-    final double u;
-    final double v;
+
 
     /**
      * <p>
@@ -20,27 +22,19 @@ public class Quantile {
      * </p>
      *
      * @param quantile The target quantile value expressed along the interval
-     *        <code>[0, 1]</code>.
-     * @param error The target error allowance expressed along the interval
-     *        <code>[0, 1]</code>.
+     *                 <code>[0, 1]</code>.
+     * @param error    The target error allowance expressed along the interval
+     *                 <code>[0, 1]</code>.
      */
     public Quantile(final double quantile, final double error) {
         this.quantile = quantile;
         this.error = error;
-        u = 2.0 * error / (1.0 - quantile);
-        v = 2.0 * error / quantile;
     }
+
 
     @Override
     public String toString() {
         return String.format("Q{q=%f, eps=%f})", quantile, error);
-    }
-
-    double delta(final double rank, final int n) {
-        if (rank <= Math.floor(quantile * n)) {
-            return u * (n - rank);
-        }
-        return v * rank;
     }
 
     public double getQuantile() {
@@ -50,4 +44,17 @@ public class Quantile {
     public double getError() {
         return error;
     }
+
+    public double f(double r, int n) {
+        if (r <= Math.floor(quantile * n)) {
+            return (2.0 * error * (n - r)) / (1.0 - quantile);
+        } else {
+            return (2.0 * error * r) / quantile;
+        }
+    }
+
 }
+
+
+
+
